exception handling to /register endpoint

+++ Profitable_Choice/src/main/java/com/steds/dao/UserWebAppDao.java
@@ -12,7 +12,7 @@ public interface UserWebAppDao {
 
     public String loginUser(User user);
 
-    public User findUser(long id);
+    public User findUser(int id);
 
     public User findUserByName(String userName);
